refactored and cleaned
