+++ src/source/raster_tile_source.ts
@@ -1,6 +1,6 @@
 import {extend, pick} from '../util/util';
 
-import {getImage} from '../util/ajax';
+import ImageRequest from '../util/imageRequest';
 import {ResourceType} from '../util/request_manager';
 import {Event, ErrorEvent, Evented} from '../util/evented';
 import loadTileJSON from './load_tilejson';
@@ -104,7 +104,7 @@ class RasterTileSource extends Evented implements Source {
 
     loadTile(tile: Tile, callback: Callback<void>) {
         const url = tile.tileID.canonical.url(this.tiles, this.map.getPixelRatio(), this.scheme);
-        tile.request = getImage(this.map._requestManager.transformRequest(url, ResourceType.Tile), (err, img, expiry) => {
+        tile.request = ImageRequest.getImage(this.map._requestManager.transformRequest(url, ResourceType.Tile), (err, img, expiry) => {
             delete tile.request;
 
             if (tile.aborted) {
