+++ src/ui/map.ts
@@ -3,14 +3,9 @@ import browser from '../util/browser';
 import DOM from '../util/dom';
 import packageJSON from '../../package.json' assert {type: 'json'};
 
-import {
-    getImage,
-    GetImageCallback,
-    getJSON,
-    processImageRequestQueue,
-    installImageQueueThrottleControlCallback,
-    removeImageQueueThrottleControlCallback
-} from '../util/ajax';
+import {getJSON} from '../util/ajax';
+import ImageRequest from '../util/imageRequest';
+import type {GetImageCallback} from '../util/imageRequest';
 
 import {RequestManager, ResourceType} from '../util/request_manager';
 import Style, {StyleSwapOptions} from '../style/style';
@@ -439,7 +434,7 @@ class Map extends Camera {
         this._clickTolerance = options.clickTolerance;
         this._pixelRatio = options.pixelRatio ?? devicePixelRatio;
 
-        this._imageQueueThrottleControlCallbackHandle = installImageQueueThrottleControlCallback(() => this.isMoving());
+        this._imageQueueThrottleControlCallbackHandle = ImageRequest.installThrottleControlCallback(() => this.isMoving());
 
         this._requestManager = new RequestManager(options.transformRequest);
 
@@ -1962,7 +1957,7 @@ class Map extends Camera {
      * @see [Add an icon to the map](https://maplibre.org/maplibre-gl-js-docs/example/add-image/)
      */
     loadImage(url: string, callback: GetImageCallback) {
-        getImage(this._requestManager.transformRequest(url, ResourceType.Image), callback);
+        ImageRequest.getImage(this._requestManager.transformRequest(url, ResourceType.Image), callback);
     }
 
     /**
@@ -2824,7 +2819,7 @@ class Map extends Camera {
         if (this.terrain) this.terrain.sourceCache.update(this.transform, this.terrain);
         this.transform.updateElevation(this.terrain);
 
-        this._imageRequestQueueDirty = processImageRequestQueue() > 0;
+        this._imageRequestQueueDirty = ImageRequest.processQueue() > 0;
 
         this._placementDirty = this.style && this.style._updatePlacement(this.painter.transform, this.showCollisionBoxes, this._fadeDuration, this._crossSourceCollisions);
 
@@ -2954,7 +2949,7 @@ class Map extends Camera {
             removeEventListener('online', this._onWindowOnline, false);
         }
 
-        removeImageQueueThrottleControlCallback(this._imageQueueThrottleControlCallbackHandle);
+        ImageRequest.removeThrottleControlCallback(this._imageQueueThrottleControlCallbackHandle);
 
         const extension = this.painter.context.gl.getExtension('WEBGL_lose_context');
         if (extension) extension.loseContext();
